Flexible spaces update

+++ pages/api/lists/[listId].ts
@@ -6,7 +6,37 @@ const listApiRoutes = async (req, res) => {
 	//mongoose code
 	await connectMongo();
 
-	if (req.method === 'PUT' && req.body.data.action === 'check') {
+	if (req.method === 'PUT' && req.body.action === 'assign-user-to-list') {
+		//look up the list
+		const list = await List.findById(req.body.listId);
+		//update the creator field, clear any custom creator values
+		list.creator = req.body.user;
+		list.customCreator = null;
+		//save
+		list.save();
+
+		res.send({
+			status: 200,
+			data: list,
+		});
+	}
+
+	if (req.method === 'PUT' && req.body.action === 'add-custom-list-creator') {
+		//look up the list
+		const list = await List.findById(req.body.listId);
+		//update the creator field, clear any collaborator creator values
+		list.customCreator = req.body.name;
+		list.creator = null;
+		//save
+		list.save();
+
+		res.send({
+			status: 200,
+			data: list,
+		});
+	}
+
+	if (req.method === 'PUT' && req.body.action === 'check') {
 		//find the user object we want to add as the resolver of this item
 		const user = await User.findOne({ email: req.body.data.userEmail });
 		if (!user) {
@@ -32,7 +62,7 @@ const listApiRoutes = async (req, res) => {
 		}
 	}
 
-	if (req.method === 'PUT' && req.body.data.action === 'uncheck') {
+	if (req.method === 'PUT' && req.body.action === 'uncheck') {
 		//find this list
 		const listItem = await ListItem.findById(req.body.data.listItemId);
 		listItem.resolvedBy = null;
@@ -41,23 +71,42 @@ const listApiRoutes = async (req, res) => {
 	}
 
 	if (req.method === 'DELETE') {
-		//Delete it from its respective list
-		const list = await List.findById(req.body.listId);
-		const newListItems = await list.items.filter(
-			(item) => item._id.toString() !== req.body.listItemId
-		);
-		list.items = newListItems;
-		await list.save();
+		if (req.body.action === 'delete-list') {
+			//Delete all of its items
+			const deletedListItems = await ListItem.deleteMany({
+				list: req.body.listId,
+			});
 
-		//Delete the list item itself
-		const deletedListIem = await ListItem.deleteOne({
-			_id: req.body.listItemId,
-		});
+			//Delete the list itself
+			const deletedList = await List.deleteOne({
+				_id: req.body.listId,
+			});
 
-		res.send({
-			status: 200,
-			data: deletedListIem,
-		});
+			res.send({
+				status: 200,
+				data: { deletedList, deletedListItems },
+			});
+		}
+
+		if (req.body.action === 'delete-list-item') {
+			//Delete it from its respective list
+			const list = await List.findById(req.body.listId);
+			const newListItems = await list.items.filter(
+				(item) => item._id.toString() !== req.body.listItemId
+			);
+			list.items = newListItems;
+			await list.save();
+
+			//Delete the list item itself
+			const deletedListIem = await ListItem.deleteOne({
+				_id: req.body.listItemId,
+			});
+
+			res.send({
+				status: 200,
+				data: deletedListIem,
+			});
+		}
 	}
 };
